ARM: tegra: resolve compilation time warnings
- Adding flag to treat warning as error.
- Handling warnings of unused variable, structures and functions,
wrong return type, wrong type comparision.

Bug 949219

Reviewed-on: http://git-master/r/104968

+++ arch/arm/mach-tegra/tegra3_usb_phy.c
@@ -1076,8 +1076,9 @@ static unsigned int utmi_phy_xcvr_setup_value(struct tegra_usb_phy *phy)
 	if (cfg->xcvr_use_fuses) {
 		val = XCVR_SETUP(tegra_fuse_readl(FUSE_USB_CALIB_0));
 		if (cfg->xcvr_use_lsb) {
-			val = min(((val & XCVR_SETUP_LSB_MASK) + cfg->xcvr_setup_offset),
-					XCVR_SETUP_LSB_MAX_VAL);
+			val = min((unsigned int) ((val & XCVR_SETUP_LSB_MASK)
+				+ cfg->xcvr_setup_offset),
+				(unsigned int) XCVR_SETUP_LSB_MAX_VAL);
 			val |= (cfg->xcvr_setup & XCVR_SETUP_MSB_MASK);
 		} else {
 			if (cfg->xcvr_setup_offset <= UTMIP_XCVR_MAX_OFFSET)
